Move params out of the helper package 🌿
